refactor(wishlist): remove WishlistContext and implement local state management for wishlist across components; eliminate unused Toast component and related toast notifications
