+++ frontend/src/App.js
@@ -52,68 +52,65 @@ import ResetPassword from './pages/pages/auth/reset-password';
 import { CartProvider } from "./contexts/CartContext";
 import Wishlist from './pages/wishlist';
 import OrderView from './pages/pages/my-account/order-view';
-import { WishlistProvider } from "./contexts/WishlistContext";
 
 
 function App() {
   return (
-    <WishlistProvider>
-      <CartProvider>
-        <Routes>
-          <Route path="/" element={<Index />}></Route>
-          <Route path="/products" element={<Products />}></Route>
-          <Route path="/shop-list" element={<ShopList />}></Route>
-          <Route path="/shop-list-left-sidebar" element={<ShopListLeftSidebar />}></Route>
-          <Route path="/shop-list-right-sidebar" element={<ShopListRightSidebar />}></Route>
-          <Route path="/shop-cart" element={<Shopcart />}></Route>
-          <Route path="/shop-checkout" element={<ShopCheckOut />}></Route>
-          <Route path="/our-store" element={<OurStore />}></Route>
-          <Route path="/brands" element={<Brands />}></Route>
-          <Route path="/compare-product" element={<CompareProduct />}></Route>
-          <Route path="/recently-viewed-product" element={<RecentlyViewed />}></Route>
-          <Route path="/aboutus" element={<AboutUS />}></Route>
-          <Route path="/user-account" element={<UserAccount />}></Route>
-          <Route path="/order-view/:id" element={<OrderView />}></Route>
-          <Route path="/user-billing" element={<UserBilling />}></Route>
-          {/* <Route path="/user-payment" element={<UserPayment />}></Route> */}
-          <Route path="/user-invoice" element={<Invoice />}></Route>
-          <Route path="/user-social" element={<UserSocial />}></Route>
-          <Route path="/user-notification" element={<Notification />}></Route>
-          <Route path="/user-setting" element={<UserSetting />}></Route>
-          <Route path="/email-cart" element={<EmailCart />}></Route>
-          <Route path="/email-confirmation" element={<EmailConfirmation />}></Route>
-          <Route path="/email-offers" element={<EmailOffers />}></Route>
-          <Route path="/email-order-success" element={<EmailOrderSuccess />}></Route>
-          <Route path="/email-gift-voucher" element={<EmailGiftVoucher />}></Route>
-          <Route path="/email-reset-password" element={<EmailResetPassword />}></Route>
-          <Route path="/email-item-review" element={<EmailItemReviw />}></Route>
-          <Route path="/blogs" element={<Blogs />}></Route>
-          <Route path="/blog-detail" element={<BlogDetail />}></Route>
-          <Route path="/blog-detail/:id" element={<BlogDetail />}></Route>
-          <Route path="/career" element={<Career />}></Route>
-          <Route path="/helpcenter" element={<Helpcenter />}></Route>
-          <Route path="/helpcenter-guides" element={<HelpcenterGuides />}></Route>
-          <Route path="/helpcenter-support" element={<HelpcenterSupport />}></Route>
-          <Route path="/helpcenter-faqs" element={<HelpcenterFaqs />}></Route>
-          <Route path="/terms" element={<Terms />}></Route>
-          <Route path="/privacy" element={<Privacy />}></Route>
-          <Route path="/login" element={<Login />}></Route>
-          <Route path="/signup" element={<Signup />}></Route>
-          <Route path="/forgot-password" element={<ForgotPassword />}></Route>
-          <Route path="/lock-screen" element={<LockScreen />}></Route>
-          <Route path="/comingsoon" element={<Comingsoon />}></Route>
-          <Route path="/maintenance" element={<Maintenance />}></Route>
-          <Route path="/error" element={<Error />}></Route>
-          <Route path="/sale" element={<Sale />}></Route>
-          <Route path="/contact" element={<Contact />}></Route>
-          <Route path="/product-detail-one" element={<ProductDetailOne />}></Route>
-          <Route path="/product-detail-one/:id" element={<ProductDetailOne />}></Route>
-          <Route path='/signup-success' element={<SignSuccess />}></Route>
-          <Route path='/reset-password/:token' element={<ResetPassword />}></Route>
-          <Route path="/wishlist" element={<Wishlist />} />
-        </Routes>
-      </CartProvider>
-    </WishlistProvider>
+    <CartProvider>
+      <Routes>
+        <Route path="/" element={<Index />}></Route>
+        <Route path="/products" element={<Products />}></Route>
+        <Route path="/shop-list" element={<ShopList />}></Route>
+        <Route path="/shop-list-left-sidebar" element={<ShopListLeftSidebar />}></Route>
+        <Route path="/shop-list-right-sidebar" element={<ShopListRightSidebar />}></Route>
+        <Route path="/shop-cart" element={<Shopcart />}></Route>
+        <Route path="/shop-checkout" element={<ShopCheckOut />}></Route>
+        <Route path="/our-store" element={<OurStore />}></Route>
+        <Route path="/brands" element={<Brands />}></Route>
+        <Route path="/compare-product" element={<CompareProduct />}></Route>
+        <Route path="/recently-viewed-product" element={<RecentlyViewed />}></Route>
+        <Route path="/aboutus" element={<AboutUS />}></Route>
+        <Route path="/user-account" element={<UserAccount />}></Route>
+        <Route path="/order-view/:id" element={<OrderView />}></Route>
+        <Route path="/user-billing" element={<UserBilling />}></Route>
+        {/* <Route path="/user-payment" element={<UserPayment />}></Route> */}
+        <Route path="/user-invoice" element={<Invoice />}></Route>
+        <Route path="/user-social" element={<UserSocial />}></Route>
+        <Route path="/user-notification" element={<Notification />}></Route>
+        <Route path="/user-setting" element={<UserSetting />}></Route>
+        <Route path="/email-cart" element={<EmailCart />}></Route>
+        <Route path="/email-confirmation" element={<EmailConfirmation />}></Route>
+        <Route path="/email-offers" element={<EmailOffers />}></Route>
+        <Route path="/email-order-success" element={<EmailOrderSuccess />}></Route>
+        <Route path="/email-gift-voucher" element={<EmailGiftVoucher />}></Route>
+        <Route path="/email-reset-password" element={<EmailResetPassword />}></Route>
+        <Route path="/email-item-review" element={<EmailItemReviw />}></Route>
+        <Route path="/blogs" element={<Blogs />}></Route>
+        <Route path="/blog-detail" element={<BlogDetail />}></Route>
+        <Route path="/blog-detail/:id" element={<BlogDetail />}></Route>
+        <Route path="/career" element={<Career />}></Route>
+        <Route path="/helpcenter" element={<Helpcenter />}></Route>
+        <Route path="/helpcenter-guides" element={<HelpcenterGuides />}></Route>
+        <Route path="/helpcenter-support" element={<HelpcenterSupport />}></Route>
+        <Route path="/helpcenter-faqs" element={<HelpcenterFaqs />}></Route>
+        <Route path="/terms" element={<Terms />}></Route>
+        <Route path="/privacy" element={<Privacy />}></Route>
+        <Route path="/login" element={<Login />}></Route>
+        <Route path="/signup" element={<Signup />}></Route>
+        <Route path="/forgot-password" element={<ForgotPassword />}></Route>
+        <Route path="/lock-screen" element={<LockScreen />}></Route>
+        <Route path="/comingsoon" element={<Comingsoon />}></Route>
+        <Route path="/maintenance" element={<Maintenance />}></Route>
+        <Route path="/error" element={<Error />}></Route>
+        <Route path="/sale" element={<Sale />}></Route>
+        <Route path="/contact" element={<Contact />}></Route>
+        <Route path="/product-detail-one" element={<ProductDetailOne />}></Route>
+        <Route path="/product-detail-one/:id" element={<ProductDetailOne />}></Route>
+        <Route path='/signup-success' element={<SignSuccess />}></Route>
+        <Route path='/reset-password/:token' element={<ResetPassword />}></Route>
+        <Route path="/wishlist" element={<Wishlist />} />
+      </Routes>
+    </CartProvider>
   );
 }
