+++ frontend/src/pages/index/index.jsx
@@ -9,8 +9,6 @@ import ScrollToTop from "../../components/scroll-to-top";
 import heroVideo from '../../assets/videos/heroVideo6.mp4';
 import { FiHeart, FiEye, FiBookmark } from '../../assets/icons/vander';
 import { AiFillHeart } from 'react-icons/ai';
-import Toast from '../../components/Toast';
-import { useWishlist } from '../../contexts/WishlistContext';
 
 export default function Index() {
   const [collections, setCollections] = useState([]);
@@ -18,8 +16,6 @@ export default function Index() {
   const [wishlist, setWishlist] = useState([]);
   const [productRatings, setProductRatings] = useState({});
   const [loading, setLoading] = useState(true);
-  const { wishlist: contextWishlist, refreshWishlist } = useWishlist();
-  const [toast, setToast] = useState({ message: '', type: 'info' });
 
   const fetchCollections = async () => {
     try {
@@ -74,20 +70,20 @@ export default function Index() {
     if (token) {
       try {
         await wishlistAPI.addToWishlistAlt({ productId: item._id });
-        await refreshWishlist();
-        setToast({ message: 'Added to wishlist!', type: 'success' });
+        setWishlist(prev => prev.includes(item._id) ? prev : [...prev, item._id]);
+        alert('Added to wishlist!');
       } catch (error) {
         console.error('Error adding to wishlist:', error);
-        setToast({ message: 'Failed to add to wishlist', type: 'error' });
+        alert('Failed to add to wishlist');
       }
     } else {
       let localWishlist = JSON.parse(localStorage.getItem('wishlist') || '[]');
       if (!localWishlist.find(p => p._id === item._id)) {
         localWishlist.push(item);
         localStorage.setItem('wishlist', JSON.stringify(localWishlist));
+        setWishlist(prev => prev.includes(item._id) ? prev : [...prev, item._id]);
       }
-      await refreshWishlist();
-      setToast({ message: 'Added to wishlist (local)!', type: 'success' });
+      alert('Added to wishlist (local)!');
     }
   };
 
@@ -240,7 +236,6 @@ export default function Index() {
 
   return (
     <>
-      <Toast message={toast.message} type={toast.type} onClose={() => setToast({ message: '', type: 'info' })} />
       <Tagline />
       <Navbar navClass="defaultscroll is-sticky tagline-height" />
